converted sum to long int

+++ sorting/act3/sort_act3_mp2525.c
@@ -45,7 +45,7 @@ int main(int argc, char **argv) {
   // Write code here
 
   double t0, t1, t2, t3, distributionTime, sortingTime, totalTime;
-  int globalSum, localSum;
+  long int globalSum, localSum;
 
   /******************************************
   * Global Sum Calculation before sorting
@@ -59,7 +59,7 @@ int main(int argc, char **argv) {
 
   // Send localsum from all ranks to 0 and reduce the sum into global sum
 
-  MPI_Reduce(&localSum, &globalSum, 1, MPI_INT, MPI_SUM, 0, MPI_COMM_WORLD);
+  MPI_Reduce(&localSum, &globalSum, 1, MPI_LONG, MPI_SUM, 0, MPI_COMM_WORLD);
 
   // Print global sum by rank 0
   if (my_rank == 0) {
@@ -234,7 +234,7 @@ int main(int argc, char **argv) {
   }
 
   // Send localsum from all ranks to 0 and reduce the sum into global sum
-  MPI_Reduce(&localSum, &globalSum, 1, MPI_INT, MPI_SUM, 0, MPI_COMM_WORLD);
+  MPI_Reduce(&localSum, &globalSum, 1, MPI_LONG, MPI_SUM, 0, MPI_COMM_WORLD);
 
   // Print global sum by rank 0
   if (my_rank == 0) {
