+++ sorting/test/sort_act2_mp2525.c
@@ -51,7 +51,7 @@ int main(int argc, char **argv) {
   printf("\n");
 
   double t0, t1, t2, t3, distributionTime, sortingTime, totalTime;
-  int globalSum, localSum;
+  long int globalSum, localSum;
 
   /******************************************
   * Global Sum Calculation before sorting
@@ -65,7 +65,7 @@ int main(int argc, char **argv) {
 
   // Send localsum from all ranks to 0 and reduce the sum into global sum
 
-  MPI_Reduce(&localSum, &globalSum, 1, MPI_INT, MPI_SUM, 0, MPI_COMM_WORLD);
+  MPI_Reduce(&localSum, &globalSum, 1, MPI_LONG, MPI_SUM, 0, MPI_COMM_WORLD);
 
   // Print global sum by rank 0
   if (my_rank == 0) {
@@ -222,7 +222,7 @@ int main(int argc, char **argv) {
   }
 
   // Send localsum from all ranks to 0 and reduce the sum into global sum
-  MPI_Reduce(&localSum, &globalSum, 1, MPI_INT, MPI_SUM, 0, MPI_COMM_WORLD);
+  MPI_Reduce(&localSum, &globalSum, 1, MPI_LONG, MPI_SUM, 0, MPI_COMM_WORLD);
 
   // Print global sum by rank 0
   if (my_rank == 0) {
